Moved double tap button to separate class and namespace
