+++ Assets/2DGamekit/Scripts/Character/MonoBehaviours/InputComponent.cs
@@ -186,93 +186,6 @@ namespace Gamekit2D
         }
 
         [Serializable]
-        public class DoubleTapButton : InputButton
-        {
-            private int m_doubleTapStep = 0;
-            private float m_doubleTapTimer = 0f;
-            private float m_doubleTapSpeed = 0.25f;
-
-            public DoubleTapButton(KeyCode key, XboxControllerButtons controllerButton)
-                : base(key, controllerButton) { }
-
-            public override void Get(bool fixedUpdateHappened, InputType inputType)
-            {
-                if (!m_Enabled)
-                {
-                    Down = false;
-                    Held = false;
-                    Up = false;
-                    return;
-                }
-                else if (!m_GettingInput)
-                {
-                    return;
-                }
-
-                if (fixedUpdateHappened)
-                {
-                    Held = m_AfterFixedUpdateHeld;
-                    Up = m_AfterFixedUpdateUp;
-
-                    if (m_doubleTapStep == 0 && m_AfterFixedUpdateDown)
-                    {
-                        m_doubleTapTimer = m_doubleTapSpeed;
-                        m_doubleTapStep = 1;
-                    }
-                    else if (m_doubleTapTimer <= 0)
-                    {
-                        m_doubleTapTimer = 0;
-                        m_doubleTapStep = 0;
-                        Down = false;
-                        Up = !Held;
-                    }
-                    else if (m_doubleTapStep == 1 && m_AfterFixedUpdateUp)
-                    {
-                        m_doubleTapTimer = m_doubleTapSpeed;
-                        m_doubleTapStep = 2;
-                    }
-                    else if (m_doubleTapStep == 2 && m_AfterFixedUpdateDown)
-                    {
-                        m_doubleTapTimer = 0;
-                        m_doubleTapStep = 3;
-                        Down = true;
-                        Up = false;
-                    }
-
-                    if (m_doubleTapTimer > 0)
-                    {
-                        m_doubleTapTimer -= Time.deltaTime;
-                    }
-
-                    m_AfterFixedUpdateDown = false;
-                    m_AfterFixedUpdateHeld = false;
-                    m_AfterFixedUpdateUp = false;
-                }
-
-                bool isDown = false;
-                bool isHeld = false;
-                bool isUp = false;
-                if (inputType == InputType.Controller)
-                {
-                    isDown = Input.GetButtonDown(k_ButtonsToName[(int)controllerButton]);
-                    isHeld = Input.GetButton(k_ButtonsToName[(int)controllerButton]);
-                    isUp = Input.GetButtonUp(k_ButtonsToName[(int)controllerButton]);
-                }
-                else if (inputType == InputType.MouseAndKeyboard)
-                {
-                    isDown = Input.GetKeyDown(key);
-                    isHeld = Input.GetKey(key);
-                    isUp = Input.GetKeyUp(key);
-                }
-
-                m_AfterFixedUpdateDown |= isDown;
-                m_AfterFixedUpdateHeld |= isHeld;
-                m_AfterFixedUpdateUp |= isUp;
-            }
-        }
-
-
-        [Serializable]
         public class InputAxis
         {
             public KeyCode positive;
