+++ Assets/Scripts/Character/MonoBehaviours.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eafe04017d5a480939cb6beb9c8a63be
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
