+++ Assets/Scripts/Character/MonoBehaviours/InputComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e971c1cf7dfb3094080ca524228547c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
