+++ Assets/Scripts/Character/StateMachineBehaviours/Player/AirborneDashSMB.cs
@@ -1,8 +1,9 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Gamekit2D;
 
-namespace Gamekit2D
+namespace Buffalo
 {
     public class AirborneDashSMB : SceneLinkedSMB<PlayerCharacter>
     {
